# apps/api/src/Adapters/In/Http/AuthController.php
<?php

declare(strict_types=1);

namespace App\Adapters\In\Http;

use App\Application\UseCases\Auth\Login\InvalidCredentials;
use App\Application\UseCases\Auth\Login\LoginCommand;
use App\Application\UseCases\Auth\Login\LoginHandler;
use App\Application\UseCases\Auth\Logout\LogoutHandler;
use App\Application\UseCases\Auth\Token\IssueAuthTokenCommand;
use App\Application\UseCases\Auth\Token\IssueAuthTokenHandler;
use App\Application\UseCases\Auth\User\CreateUser\CreateUserAlreadyExists;
use App\Application\UseCases\Auth\User\CreateUser\CreateUserCommand;
use App\Application\UseCases\Auth\User\CreateUser\CreateUserHandler;
use App\Application\UseCases\Auth\User\CreateUser\CreateUserValidationException;
use App\Application\UseCases\Auth\User\DeleteUser\DeleteUserByEmailCommand;
use App\Application\UseCases\Auth\User\DeleteUser\DeleteUserByEmailHandler;
use App\Application\UseCases\Auth\User\DeleteUser\DeleteUserNotFound;
use App\Application\UseCases\Auth\User\DeleteUser\DeleteUserValidationException;
use App\Application\UseCases\Auth\User\GetCurrentUserHandler;
use App\Application\UseCases\Auth\User\UpdateCurrentUser\UpdateCurrentUserCommand;
use App\Application\UseCases\Auth\User\UpdateCurrentUser\UpdateCurrentUserHandler;
use App\Application\UseCases\Auth\User\UpdateCurrentUser\UpdateCurrentUserNotFound;
use App\Application\UseCases\Auth\User\UpdateCurrentUser\UpdateCurrentUserUnauthenticated;
use App\Application\UseCases\Auth\User\UpdateCurrentUser\UpdateCurrentUserValidationException;
use App\Application\Ports\AuthSessionManager;
use App\Domain\Model\AuthUser;
use Symfony\Component\HttpFoundation\JsonResponse;
use Symfony\Component\HttpFoundation\Request;
use Symfony\Component\HttpFoundation\Response;
use Symfony\Component\Routing\Attribute\Route;

final class AuthController
{
    public function __construct(
        private readonly LoginHandler $loginHandler,
        private readonly IssueAuthTokenHandler $issueAuthTokenHandler,
        private readonly GetCurrentUserHandler $getCurrentUserHandler,
        private readonly LogoutHandler $logoutHandler,
        private readonly CreateUserHandler $createUserHandler,
        private readonly DeleteUserByEmailHandler $deleteUserByEmailHandler,
        private readonly UpdateCurrentUserHandler $updateCurrentUserHandler,
        private readonly AuthSessionManager $authSession,
    ) {
    }

    #[Route('/api/auth/login', name: 'api_auth_login', methods: ['POST'])]
    public function login(Request $request): JsonResponse
    {
        $credentials = $this->parseCredentialsPayload($request);
        if ($credentials instanceof JsonResponse) {
            return $credentials;
        }

        try {
            $user = $this->loginHandler->handle(new LoginCommand($credentials['email'], $credentials['password']));
        } catch (InvalidCredentials) {
            return new JsonResponse(['error' => 'Invalid credentials.'], Response::HTTP_UNAUTHORIZED);
        }

        return new JsonResponse(['user' => $this->userPayload($user)], Response::HTTP_OK);
    }

    #[Route('/api/auth/token', name: 'api_auth_token', methods: ['POST'])]
    public function token(Request $request): JsonResponse
    {
        $credentials = $this->parseCredentialsPayload($request);
        if ($credentials instanceof JsonResponse) {
            return $credentials;
        }

        try {
            $result = $this->issueAuthTokenHandler->handle(new IssueAuthTokenCommand(
                email: $credentials['email'],
                password: $credentials['password'],
            ));
        } catch (InvalidCredentials) {
            return new JsonResponse(['error' => 'Invalid credentials.'], Response::HTTP_UNAUTHORIZED);
        }

        return new JsonResponse([
            'access_token' => $result->token,
            'token_type' => 'Bearer',
            'expires_at' => $result->expiresAt->format(\DateTimeInterface::ATOM),
            'user' => $this->userPayload($result->user),
        ], Response::HTTP_OK);
    }

    #[Route('/api/auth/me', name: 'api_auth_me', methods: ['GET'])]
    public function me(): JsonResponse
    {
        $user = $this->getCurrentUserHandler->handle();
        if (null === $user) {
            return new JsonResponse(['error' => 'Unauthenticated.'], Response::HTTP_UNAUTHORIZED);
        }

        return new JsonResponse(['user' => $this->userPayload($user)], Response::HTTP_OK);
    }

    #[Route('/api/auth/me', name: 'api_auth_me_update', methods: ['PUT'])]
    public function updateCurrentUser(Request $request): JsonResponse
    {
        $payload = json_decode($request->getContent(), true);
        if (!is_array($payload)) {
            return new JsonResponse(['error' => 'Invalid JSON body.'], Response::HTTP_BAD_REQUEST);
        }

        $name = $payload['name'] ?? null;
        $lastname = $payload['lastname'] ?? null;
        $password = $payload['password'] ?? null;

        if (!is_string($name) || !is_string($lastname) || (null !== $password && !is_string($password))) {
            return new JsonResponse(['error' => 'name and lastname are required, password must be a string when provided.'], Response::HTTP_BAD_REQUEST);
        }

        try {
            $user = $this->updateCurrentUserHandler->handle(new UpdateCurrentUserCommand(
                name: $name,
                lastname: $lastname,
                password: $password,
            ));
        } catch (UpdateCurrentUserUnauthenticated $exception) {
            return new JsonResponse(['error' => $exception->getMessage()], Response::HTTP_UNAUTHORIZED);
        } catch (UpdateCurrentUserValidationException $exception) {
            return new JsonResponse(['error' => $exception->getMessage()], Response::HTTP_BAD_REQUEST);
        } catch (UpdateCurrentUserNotFound $exception) {
            return new JsonResponse(['error' => $exception->getMessage()], Response::HTTP_NOT_FOUND);
        }

        return new JsonResponse(['user' => $this->userPayload($user)], Response::HTTP_OK);
    }

    #[Route('/api/auth/logout', name: 'api_auth_logout', methods: ['POST'])]
    public function logout(): JsonResponse
    {
        $this->logoutHandler->handle();

        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
    }

    #[Route('/api/auth/users', name: 'api_auth_users_create', methods: ['POST'])]
    public function createUser(Request $request): JsonResponse
    {
        if (null === $this->authSession->getAuthenticatedUserId()) {
            return new JsonResponse(['error' => 'Unauthenticated.'], Response::HTTP_UNAUTHORIZED);
        }

        $payload = json_decode($request->getContent(), true);
        if (!is_array($payload)) {
            return new JsonResponse(['error' => 'Invalid JSON body.'], Response::HTTP_BAD_REQUEST);
        }

        $email = $payload['email'] ?? null;
        $name = $payload['name'] ?? null;
        $lastname = $payload['lastname'] ?? null;
        $password = $payload['password'] ?? null;
        if (!is_string($email) || !is_string($name) || !is_string($lastname) || !is_string($password)) {
            return new JsonResponse(['error' => 'email, name, lastname and password are required.'], Response::HTTP_BAD_REQUEST);
        }

        try {
            $user = $this->createUserHandler->handle(new CreateUserCommand($email, $name, $lastname, $password));
        } catch (CreateUserValidationException $exception) {
            return new JsonResponse(['error' => $exception->getMessage()], Response::HTTP_BAD_REQUEST);
        } catch (CreateUserAlreadyExists $exception) {
            return new JsonResponse(['error' => $exception->getMessage()], Response::HTTP_CONFLICT);
        }

        return new JsonResponse(['user' => $this->userPayload($user)], Response::HTTP_CREATED);
    }

    #[Route('/api/auth/users', name: 'api_auth_users_delete', methods: ['DELETE'])]
    public function deleteUser(Request $request): JsonResponse
    {
        if (null === $this->authSession->getAuthenticatedUserId()) {
            return new JsonResponse(['error' => 'Unauthenticated.'], Response::HTTP_UNAUTHORIZED);
        }

        $payload = json_decode($request->getContent(), true);
        if (!is_array($payload)) {
            return new JsonResponse(['error' => 'Invalid JSON body.'], Response::HTTP_BAD_REQUEST);
        }

        $email = $payload['email'] ?? null;
        if (!is_string($email)) {
            return new JsonResponse(['error' => 'email is required.'], Response::HTTP_BAD_REQUEST);
        }

        try {
            $this->deleteUserByEmailHandler->handle(new DeleteUserByEmailCommand($email));
        } catch (DeleteUserValidationException $exception) {
            return new JsonResponse(['error' => $exception->getMessage()], Response::HTTP_BAD_REQUEST);
        } catch (DeleteUserNotFound $exception) {
            return new JsonResponse(['error' => $exception->getMessage()], Response::HTTP_NOT_FOUND);
        }

        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
    }

    /**
     * @return array{id:int,email:string,name:string,lastname:string}
     */
    private function userPayload(AuthUser $user): array
    {
        return [
            'id' => $user->id,
            'email' => $user->email,
            'name' => $user->name,
            'lastname' => $user->lastname,
        ];
    }

    /**
     * @return array{email:string,password:string}|JsonResponse
     */
    private function parseCredentialsPayload(Request $request): array|JsonResponse
    {
        $payload = json_decode($request->getContent(), true);
        if (!is_array($payload)) {
            return new JsonResponse(['error' => 'Invalid JSON body.'], Response::HTTP_BAD_REQUEST);
        }

        $email = $payload['email'] ?? null;
        $password = $payload['password'] ?? null;

        if (!is_string($email) || '' === trim($email) || !is_string($password) || '' === $password) {
            return new JsonResponse(['error' => 'email and password are required.'], Response::HTTP_BAD_REQUEST);
        }

        return [
            'email' => $email,
            'password' => $password,
        ];
    }
}