# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Wine App API
  version: 0.1.0
  description: Current API endpoints for health, authentication, wines, reviews, stats, and wine photos. Web clients continue to use the Symfony session cookie; mobile clients should prefer bearer JWT auth.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://dev-api.example.com
    description: Shared development environment placeholder
  - url: https://staging-api.example.com
    description: Staging environment placeholder
  - url: https://api.example.com
    description: Production environment placeholder
paths:
...
  /api/auth/login:
    post:
      summary: Login with email and password (session cookie)
      operationId: postAuthLogin
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully
          headers:
            Set-Cookie:
              description: Symfony session cookie issued for web clients
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/token:
    post:
      summary: Issue a bearer token from email and password
      operationId: postAuthToken
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthTokenRequest'
      responses:
        '200':
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      summary: Get current authenticated user
      operationId: getAuthMe
      tags: [Auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update current authenticated user
      operationId: putAuthMe
      tags: [Auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentUserRequest'
      responses:
        '200':
          description: Updated authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authenticated user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      summary: Logout current session
      operationId: postAuthLogout
      tags: [Auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '204':
          description: Logged out (idempotent)
  /api/auth/users:
    post:
      summary: Create auth user
      operationId: postAuthUsersCreate
      tags: [Auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete user by email
      operationId: deleteAuthUsersByEmail
      tags: [Auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAuthUserRequest'
      responses:
        '204':
          description: User deleted
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: PHPSESSID
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: adriamuixi@gmail.com
        password:
          type: string
          format: password
          example: wine1234
    CreateAuthUserRequest:
      type: object
      required: [email, name, lastname, password]
      properties:
        email:
          type: string
          format: email
          example: adriamuixi@gmail.com
        name:
          type: string
          example: Adria
        lastname:
          type: string
          example: Muixi
        password:
          type: string
          format: password
          example: wine1234
    DeleteAuthUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: adriamuixi@gmail.com
    UpdateCurrentUserRequest:
      type: object
      required: [name, lastname]
      properties:
        name:
          type: string
          example: Adria
        lastname:
          type: string
          example: Muixi
        password:
          type: string
          nullable: true
          format: password
          example: newSecret123
    AuthUser:
      type: object
      required: [id, email, name, lastname]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: adriamuixi@gmail.com
        name:
          type: string
          example: Adria
        lastname:
          type: string
          example: Muixi
    AuthUserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
    AuthTokenRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: adriamuixi@gmail.com
        password:
          type: string
          format: password
          example: wine1234
    AuthTokenResponse:
      type: object
      required: [access_token, token_type, expires_at, user]
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum: [Bearer]
          example: Bearer
        expires_at:
          type: string
          format: date-time
          example: '2030-01-01T00:00:00+00:00'
        user:
          $ref: '#/components/schemas/AuthUser'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Invalid credentials.
...